+++ src/utils/index.ts
@@ -1,4 +1,5 @@
 export * from './defineProperty';
 export * from './log';
 export * from './simpleCombineReducers';
+export * from './transformDeep';
 export * from './utils';